remove test_util.go
